float32 SOUNDNESS FIX (code review #2): GoFloat32 made an ABSTRACT smart-constructor type — close the raw-literal injection hole, zero new axioms
A code review found GoFloat32 was UNSOUND: defined as the transparent alias
`GoFloat32 := float`, the type system let a NON-binary32-representable `float`
literal flow straight into a GoFloat32 position with no rounding. Exploit:
`f64_of_f32 16777217` stays 16777217 in Rocq, but Go rounds `float32(16777217)`
to 16777216 — a Rocq/Go DISAGREEMENT on widening that licenses unsound proofs.
(The emitted Go was already float32-correct; the hole was purely Rocq-level.)

FIX — GoFloat32 is now an ABSTRACT record carrying an UNFORGEABLE provenance proof:

  Record GoFloat32 := mkF32 { f32val : float ; f32ok : exists a, f32val = f32_round a }.

f32ok witnesses the carrier is in the IMAGE of f32_round (binary32-representable).
`mkF32 16777217 _` is unconstructable (its obligation `exists a, 16777217 =
f32_round a` is false), so every inhabitant MUST enter through a rounding smart
constructor (f32_of_f64 / f32_lit / the arith ops, which route their SpecFloat
result back through f32_of_f64). Widening `f64_of_f32 := f32val` (identity) is
therefore SOUND by construction, and the raw injection no longer typechecks.

ZERO new axioms. The provenance proofs are `eq_refl` (the carrier is *literally*
f32_round a) — this deliberately SIDESTEPS the hard `binary_round_aux` rounding-
idempotence proof by carrying provenance instead of the fixpoint equation. No
Flocq, no third-party deps, no Admitted. `Print Assumptions f32_widen_sound`
(real Fido context) = Rocq's own int/float primitives ONLY (PrimInt63.*, float,
opp/eqb/div/ldshiftexp/frshiftexp/…) — exactly the trust base rule 3 permits.

EXTRACTION UNCHANGED. GoFloat32 erases to native Go `float32` and mkF32/f32val/
f32_round to identity (suppressed — they only appear inside by-name-lowered ops),
the same GoU8-style wrapper erasure (is_erased_record_typename + go_prim_type_table
+ identity ctor/proj rules in the plugin). Golden output is byte-identical; the
only Go delta is benign `float32(...)` casts on the demo literals (now via f32_lit).

Machine-checked regression `f32_widen_sound`: `widen64 (f32_lit 16777217) =
16777216`, matching Go. All existing float32 witnesses (arith/cmp, the 2^24+1
rounding contrast) re-proved through the abstract boundary. float32 literals now
enter via f32_lit (rounds at the Rocq boundary), closing the "literals beyond the
demo" item. SPEC_CONFORMANCE float32 → "✓ DONE & SOUND"; PROGRESS ladder updated.

+++ main.go
@@ -120,7 +120,7 @@ func F32_combine(a float32, b float32, c float32) float32 {
 }
 
 func F32_demo() {
-	println(F32_combine(1.5, 2.25, 2.0))
+	println(F32_combine(float32(1.5), float32(2.25), float32(2.0)))
 }
 
 func I64_of_narrow_demo() {
@@ -153,7 +153,7 @@ func Fconst_demo() {
 }
 
 func F32_cmp_demo() {
-	println(F32_combine(1.5, 0.0, 2.0) < F32_combine(5.0, 0.0, 2.0), F32_combine(5.0, 0.0, 2.0) >= F32_combine(1.5, 0.0, 2.0), F32_combine(1.5, 0.0, 2.0) != F32_combine(5.0, 0.0, 2.0))
+	println(F32_combine(float32(1.5), float32(0.0), float32(2.0)) < F32_combine(float32(5.0), float32(0.0), float32(2.0)), F32_combine(float32(5.0), float32(0.0), float32(2.0)) >= F32_combine(float32(1.5), float32(0.0), float32(2.0)), F32_combine(float32(1.5), float32(0.0), float32(2.0)) != F32_combine(float32(5.0), float32(0.0), float32(2.0)))
 }
 
 func Trunc64(x float64) int64 {
